# contracts/public.openapi.yaml
openapi: 3.1.0
info:
  title: redAI Personal owner API
  version: 1.0.0-draft.1
  description: Normative interface specification, not an implemented service. Domain
    authorization remains mandatory.
servers:
- url: /api/v1
security:
- OwnerSession: []
paths:
  /auth/login:
    post:
      operationId: login
      summary: login
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: ./schemas/api.schema.json#/$defs/Session
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/api.schema.json#/$defs/LoginRequest
  /auth/logout:
    post:
      operationId: logout
      summary: logout
      responses:
        '204':
          description: No content
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema: &id001
          type: string
          minLength: 32
  /auth/session:
    get:
      operationId: getSession
      summary: getSession
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: ./schemas/api.schema.json#/$defs/Session
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
...
  /auth/sessions:
    get:
      operationId: listOwnerSessions
      summary: listOwnerSessions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: ./schemas/api.schema.json#/$defs/SessionPage
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
      - name: cursor
        in: query
        schema:
          type: string
          minLength: 1
          maxLength: 2048
  /auth/sessions/{session_id}/revoke:
    post:
      operationId: revokeOwnerSession
      summary: revokeOwnerSession
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: *id004
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/api.schema.json#/$defs/Reason
  /auth/password:
    post:
      operationId: changeOwnerPassword
      summary: changeOwnerPassword
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: *id004
        default:
          description: Structured error
          content:
            application/json:
              schema:
                $ref: ./schemas/common.schema.json#/$defs/Error
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema: *id001
      - name: Idempotency-Key
        in: header
        required: true
        schema: *id002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/api.schema.json#/$defs/PasswordChange
...
components:
  securitySchemes:
    OwnerSession:
      type: apiKey
      in: cookie
      name: __Host-redai_session